fixed bugs

+++ src/main/java/pl/kurs/service/BikeService.java
@@ -6,7 +6,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.kurs.entity.Bike;
 import pl.kurs.exception.DataNotFoundException;
+import pl.kurs.exception.EntityCannotBeDeleteException;
 import pl.kurs.repository.BikeRepository;
+import pl.kurs.repository.ReservationRepository;
 
 import java.util.List;
 
@@ -14,6 +16,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BikeService {
     private final BikeRepository bikeRepository;
+    private final ReservationRepository reservationRepository;
 
     @Transactional
     public Bike save(Bike bike) {
@@ -22,17 +25,22 @@ public class BikeService {
 
     public Bike getBikeById(Long id) {
         return bikeRepository.findById(id)
-                .orElseThrow(() -> new DataNotFoundException("Bike not found with id: " + id));
+                .orElseThrow(() -> new DataNotFoundException(STR."Bike not found with id: \{id}"));
     }
 
+    @Transactional
     public void deleteById(Long id) {
-        bikeRepository.deleteById(id);
+        if (bikeRepository.existsById(id)) {
+            if (reservationRepository.existsByBikeId(id)) {
+             throw new EntityCannotBeDeleteException(STR."Cannot delete bike with id: \{id} - the bike has active reservations");
+            }
+            bikeRepository.deleteById(id);
+        }
     }
 
     @Transactional
     public Bike updateBike(Bike bike) {
-        Bike bikeToUpdate = bikeRepository.findById(bike.getId())
-                .orElseThrow(() -> new DataNotFoundException("Bike with id: " + bike.getId() + " not found"));
+        Bike bikeToUpdate = getBikeByIdForUpdate(bike.getId());
         BeanUtils.copyProperties(bike, bikeToUpdate);
         return bikeRepository.save(bikeToUpdate);
     }
@@ -44,4 +52,11 @@ public class BikeService {
         }
         return bikes;
     }
+
+    public Bike getBikeByIdForUpdate(Long id) {
+        return bikeRepository.findByIdForUpdate(id)
+                .orElseThrow(() -> new DataNotFoundException(STR."Bike not found with id: \{id}"));
+    }
+
+
 }
